# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Education Consultancy Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # LEADS
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: agentId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentsListResponse"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
...
  /universities:
    get:
      operationId: listUniversities
      tags: [universities]
      summary: List universities
      parameters:
        - name: country
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of universities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniversitiesListResponse"
    post:
      operationId: createUniversity
      tags: [universities]
      summary: Create university
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUniversityBody"
      responses:
        "201":
          description: Created university
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/University"
  /universities/{id}:
    get:
      operationId: getUniversity
      tags: [universities]
      summary: Get university by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: University
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/University"
    patch:
      operationId: updateUniversity
      tags: [universities]
      summary: Update university
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUniversityBody"
      responses:
        "200":
          description: Updated university
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/University"
    delete:
      operationId: deleteUniversity
      tags: [universities]
      summary: Delete university
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /programs:
    get:
      operationId: listPrograms
      tags: [universities]
      summary: List programs
      parameters:
        - name: universityId
          in: query
          schema:
            type: integer
        - name: country
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of programs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramsListResponse"
    post:
      operationId: createProgram
      tags: [universities]
      summary: Create program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramBody"
      responses:
        "201":
          description: Created program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [universities]
      summary: Get program by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    patch:
      operationId: updateProgram
      tags: [universities]
      summary: Update program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgramBody"
      responses:
        "200":
          description: Updated program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    delete:
      operationId: deleteProgram
      tags: [universities]
      summary: Delete program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # FINANCE
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - limit
        - totalPages
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        replitId:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        language:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - role
        - language
        - isActive
        - createdAt
    CreateUserBody:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        language:
          type: string
      required:
        - email
        - firstName
        - lastName
        - role
        - language
    UpdateUserBody:
      type: object
      properties:
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    UsersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
...
    Student:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        passportNumber:
          type: ["string", "null"]
        passportExpiry:
          type: ["string", "null"]
        status:
          type: string
        agentId:
          type: ["integer", "null"]
        highSchool:
          type: ["string", "null"]
        graduationYear:
          type: ["integer", "null"]
        gpa:
          type: ["string", "null"]
        languageScore:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - status
        - createdAt
        - updatedAt
    CreateStudentBody:
      type: object
      properties:
        userId:
          type: ["integer", "null"]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        passportNumber:
          type: ["string", "null"]
        passportExpiry:
          type: ["string", "null"]
        status:
          type: string
        agentId:
          type: ["integer", "null"]
        highSchool:
          type: ["string", "null"]
        graduationYear:
          type: ["integer", "null"]
        gpa:
          type: ["string", "null"]
        languageScore:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - firstName
        - lastName
        - status
    UpdateStudentBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        passportNumber:
          type: ["string", "null"]
        passportExpiry:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        agentId:
          type: ["integer", "null"]
        highSchool:
          type: ["string", "null"]
        graduationYear:
          type: ["integer", "null"]
        gpa:
          type: ["string", "null"]
        languageScore:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    StudentsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
...
    University:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        city:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        ranking:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - country
        - isActive
        - createdAt
    CreateUniversityBody:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        city:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        ranking:
          type: ["integer", "null"]
        isActive:
          type: boolean
      required:
        - name
        - country
        - isActive
    UpdateUniversityBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        ranking:
          type: ["integer", "null"]
        isActive:
          type: ["boolean", "null"]
    UniversitiesListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/University"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    Program:
      type: object
      properties:
        id:
          type: integer
        universityId:
          type: integer
        name:
          type: string
        degree:
          type: ["string", "null"]
        field:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        tuitionFee:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        scholarship:
          type: ["number", "null"]
        intakes:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        commissionRate:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - universityId
        - name
        - isActive
        - createdAt
    CreateProgramBody:
      type: object
      properties:
        universityId:
          type: integer
        name:
          type: string
        degree:
          type: ["string", "null"]
        field:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        tuitionFee:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        scholarship:
          type: ["number", "null"]
        intakes:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        commissionRate:
          type: ["number", "null"]
        isActive:
          type: boolean
      required:
        - universityId
        - name
        - isActive
    UpdateProgramBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        degree:
          type: ["string", "null"]
        field:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        tuitionFee:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        scholarship:
          type: ["number", "null"]
        intakes:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        commissionRate:
          type: ["number", "null"]
        isActive:
          type: ["boolean", "null"]
    ProgramsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Program"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
...