# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Education Consultancy Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  /** @nullable */
  userId?: number | null;
  firstName: string;
  lastName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  passportNumber?: string | null;
  /** @nullable */
  passportExpiry?: string | null;
  status: string;
  /** @nullable */
  agentId?: number | null;
  /** @nullable */
  highSchool?: string | null;
  /** @nullable */
  graduationYear?: number | null;
  /** @nullable */
  gpa?: string | null;
  /** @nullable */
  languageScore?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}