# lib/api-zod/src/generated/types/updateProgramBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Education Consultancy Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProgramBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  degree?: string | null;
  /** @nullable */
  field?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  tuitionFee?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  scholarship?: number | null;
  /** @nullable */
  intakes?: string | null;
  /** @nullable */
  requirements?: string | null;
  /** @nullable */
  commissionRate?: number | null;
  /** @nullable */
  isActive?: boolean | null;
}