# lib/api-zod/src/generated/types/updateStudentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Education Consultancy Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStudentBody {
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  passportNumber?: string | null;
  /** @nullable */
  passportExpiry?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  agentId?: number | null;
  /** @nullable */
  highSchool?: string | null;
  /** @nullable */
  graduationYear?: number | null;
  /** @nullable */
  gpa?: string | null;
  /** @nullable */
  languageScore?: string | null;
  /** @nullable */
  notes?: string | null;
}